+++ Dockerfile
@@ -1,9 +1,10 @@
-FROM golang:1.18 as base
+FROM golang:alpine3.17 as build
+WORKDIR /opt/app
+COPY . .
+RUN go build -o whosbest cmd/whosbest/main.go
 
-FROM base as dev
-
-RUN curl -sSfL https://raw.githubusercontent.com/cosmtrek/air/master/install.sh | sh -s -- -b $(go env GOPATH)/bin
-
-WORKDIR /opt/app/api
-
-CMD ["air"]
+FROM alpine:latest
+WORKDIR /opt/app
+COPY --from=build /opt/app/whosbest .
+COPY --from=build /opt/app/.env.example .env
+ENTRYPOINT ["./whosbest"]
