+++ internal/graphql/types/competitor.go
