+++ internal/helpers/utils.go
@@ -8,8 +8,8 @@ import (
 	"net/http"
 	"regexp"
 
-	"github.com/joaovicdsantos/whosbest-api/app/models"
-	"github.com/joaovicdsantos/whosbest-api/app/services"
+	"github.com/joaovicdsantos/whosbest-api/internal/models"
+	"github.com/joaovicdsantos/whosbest-api/internal/services"
 )
 
 func GetUrlParam(url string, re *regexp.Regexp) interface{} {
