+++ internal/models/competitor.go
